+++ scripts/generic/summon-long-rest.js
@@ -1,0 +1,5 @@
+if (workflow?.summonedCreatures) {
+    for (let token of workflow.summonedCreatures) {
+        await token.actor.longRest({dialog: false, chat: false, newDay: true});
+    }
+}